refactor(setup): extract wizard step 3 (connectivity check)
Moves WizardService.check_ssh_port and the wizard_check_ports route
into a new setup/wizard/step3_connectivity.py module (Step3ConnectivityMixin
+ step3_router), mirroring the Task 2 strategy-routes extraction.

Retargets monkeypatch/patch targets in test_routes.py::TestCheckPorts and
test_wizard_routes.py::TestWizardCheckPorts from
opencloudtouch.setup.wizard_service.check_ssh_port to
opencloudtouch.setup.wizard.step3_connectivity.check_ssh_port, since the
free function is now imported into the new module's namespace rather than
wizard_service's.
